Move TypeTypology to its own file
